requirements.txt
streamlit
langchain
langchain-ollama
langchain-community
langgraph>=0.0.40
python-dotenv>=1.0.0
openai>=1.0.0
pypdf2>=3.0.0
python-docx>=0.8.11
pytesseract>=0.3.10
pillow>=9.0.0
pandas>=1.5.0
numpy>=1.24.0
requests>=2.28.0
typing-extensions>=4.5.0
langchain-docling
docling
